+++ docs/changelog.rst
@@ -6,6 +6,8 @@ Change Log
 
 **Improvements**
     - removed the python.org tests from the deployed package to save space.
+    - the new fab task `install` which installs script wrappers
+      https://github.com/nexiles/nexiles.jython/issues/3
 
 **Bug Fixes**
     - Fix a bug where distutils would break.